# ior fission + benchmark to show iand fission is unprofitable