Speed up `ior`:

before
```
Benchmark                              (thatDensity)  (thisDensity)  Mode  Cnt     Score    Error  Units
BitmapContainerOrBenchmark.stable                0.1            0.1  avgt    5  1022.169 ± 42.764  ns/op
BitmapContainerOrBenchmark.stable                0.1            0.5  avgt    5  1023.257 ± 63.641  ns/op
BitmapContainerOrBenchmark.stable                0.5            0.1  avgt    5  1016.649 ± 41.469  ns/op
BitmapContainerOrBenchmark.stable                0.5            0.5  avgt    5  1020.667 ± 66.231  ns/op
BitmapContainerOrBenchmark.tendToZero            0.1            0.1  avgt    5   411.551 ± 13.539  ns/op
BitmapContainerOrBenchmark.tendToZero            0.1            0.5  avgt    5   410.931 ± 21.078  ns/op
BitmapContainerOrBenchmark.tendToZero            0.5            0.1  avgt    5   408.868 ±  1.472  ns/op
BitmapContainerOrBenchmark.tendToZero            0.5            0.5  avgt    5   411.139 ± 17.028  ns/op
```

after
```
Benchmark                              (thatDensity)  (thisDensity)  Mode  Cnt    Score    Error  Units
BitmapContainerOrBenchmark.stable                0.1            0.1  avgt    5  852.513 ± 25.221  ns/op
BitmapContainerOrBenchmark.stable                0.1            0.5  avgt    5  860.193 ± 13.692  ns/op
BitmapContainerOrBenchmark.stable                0.5            0.1  avgt    5  860.506 ± 11.027  ns/op
BitmapContainerOrBenchmark.stable                0.5            0.5  avgt    5  865.456 ± 43.538  ns/op
BitmapContainerOrBenchmark.tendToZero            0.1            0.1  avgt    5  326.623 ±  9.777  ns/op
BitmapContainerOrBenchmark.tendToZero            0.1            0.5  avgt    5  332.203 ± 18.366  ns/op
BitmapContainerOrBenchmark.tendToZero            0.5            0.1  avgt    5  327.621 ±  9.179  ns/op
BitmapContainerOrBenchmark.tendToZero            0.5            0.5  avgt    5  324.546 ±  1.616  ns/op
```

Show the same thing doesn't work for `iand` whenever an array is materialised (this is what should be optimised for)

before
```
Benchmark                               (thatDensity)  (thisDensity)  Mode  Cnt     Score     Error  Units
BitmapContainerAndBenchmark.stable                0.1            0.1  avgt    5  2768.557 ±  27.729  ns/op
BitmapContainerAndBenchmark.stable                0.1            0.5  avgt    5  4176.572 ±  36.572  ns/op
BitmapContainerAndBenchmark.stable                0.5            0.1  avgt    5  4697.707 ± 153.624  ns/op
BitmapContainerAndBenchmark.stable                0.5            0.5  avgt    5   898.776 ±  16.378  ns/op
BitmapContainerAndBenchmark.tendToZero            0.1            0.1  avgt    5  1771.341 ±  18.681  ns/op
BitmapContainerAndBenchmark.tendToZero            0.1            0.5  avgt    5  3959.568 ±  57.410  ns/op
BitmapContainerAndBenchmark.tendToZero            0.5            0.1  avgt    5  3957.309 ± 103.017  ns/op
BitmapContainerAndBenchmark.tendToZero            0.5            0.5  avgt    5   420.588 ±   1.777  ns/op
```

after (deleted code)
```
Benchmark                               (thatDensity)  (thisDensity)  Mode  Cnt     Score    Error  Units
BitmapContainerAndBenchmark.stable                0.1            0.1  avgt    5  2192.006 ± 25.172  ns/op
BitmapContainerAndBenchmark.stable                0.1            0.5  avgt    5  4627.247 ± 28.953  ns/op
BitmapContainerAndBenchmark.stable                0.5            0.1  avgt    5  5786.044 ± 45.021  ns/op
BitmapContainerAndBenchmark.stable                0.5            0.5  avgt    5   907.093 ± 15.054  ns/op
BitmapContainerAndBenchmark.tendToZero            0.1            0.1  avgt    5  1610.726 ± 35.906  ns/op
BitmapContainerAndBenchmark.tendToZero            0.1            0.5  avgt    5  4080.217 ± 67.021  ns/op
BitmapContainerAndBenchmark.tendToZero            0.5            0.1  avgt    5  4282.733 ± 75.307  ns/op
BitmapContainerAndBenchmark.tendToZero            0.5            0.5  avgt    5   328.276 ± 19.706  ns/op
```