@richardstartin 2022-02-27
> @richardstartin according to the benchmark result you gave in the pr description, the test `Score` of `BitmapContainerOrBenchmark ` after pr becomes lower.
> 
> Is it true that the lower the score, the better the performance?

The units are ns/op, so lower is better